+++ frontend/src/components/Kanban/Kanban.js
@@ -1,98 +1,16 @@
-import Todo from "../Todo/Todo"
-
-import * as React from 'react';
-import PropTypes from 'prop-types';
-import AppBar from '@mui/material/AppBar';
-import Toolbar from '@mui/material/Toolbar';
-import Typography from '@mui/material/Typography';
-import CssBaseline from '@mui/material/CssBaseline';
-import useScrollTrigger from '@mui/material/useScrollTrigger';
-import Box from '@mui/material/Box';
-import Container from '@mui/material/Container';
-import Fab from '@mui/material/Fab';
-import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp';
-import Zoom from '@mui/material/Zoom';
+import Container from "@mui/material/Container";
 import TodoList from "../TodoList/TodoList";
 import {IconButton} from "@mui/material";
-import AddIcon from '@mui/icons-material/Add';
-import NewTodo from '../NewTodo/NewTodo'
-import {useRef} from "react";
+import AddIcon from "@mui/icons-material/Add";
+import NewTodo from "../NewTodo/NewTodo";
+import * as React from "react";
 
 
-function ScrollTop(props) {
-    const { children, window } = props;
-    // Note that you normally won't need to set the window ref as useScrollTrigger
-    // will default to window.
-    // This is only being set here because the demo is in an iframe.
-    const trigger = useScrollTrigger({
-        target: window ? window() : undefined,
-        disableHysteresis: true,
-        threshold: 100,
-    });
-
-    const handleClick = (event) => {
-        const anchor = (event.target.ownerDocument || document).querySelector(
-            '#back-to-top-anchor',
-        );
-
-        if (anchor) {
-            anchor.scrollIntoView({
-                behavior: 'smooth',
-                block: 'center',
-            });
-        }
-    };
-
-    return (
-        <Zoom in={trigger}>
-            <Box
-                onClick={handleClick}
-                role="presentation"
-                sx={{  position: 'fixed', bottom: 16, right: 16 }}
-            >
-                {children}
-            </Box>
-        </Zoom>
-    );
-}
-
-ScrollTop.propTypes = {
-    children: PropTypes.element.isRequired,
-    /**
-     * Injected by the documentation to work in an iframe.
-     * You won't need it on your project.
-     */
-    window: PropTypes.func,
-};
-
-
-export default function BackToTop(props) {
-
-    const newTodoRef = useRef();
-    const anchorRef = useRef(null);
-    return (
-        <React.Fragment>
-            <CssBaseline />
-            <AppBar ref={anchorRef}>
-                <Toolbar>
-                    <Typography variant="h6" component="div">
-                        Scroll to see button
-                    </Typography>
-                </Toolbar>
-            </AppBar>
-            <Toolbar id="back-to-top-anchor" />
-            <Container maxWidth={false}>
-                <TodoList/>
-            </Container>
-            <IconButton color="primary" onClick={()=>{newTodoRef.current.handleClick(anchorRef)}}>
-                <AddIcon />
-            </IconButton>
-            <NewTodo ref={newTodoRef} />
-            <ScrollTop {...props}>
-                <Fab color="secondary" size="small" aria-label="scroll back to top">
-                    <KeyboardArrowUpIcon />
-                </Fab>
-            </ScrollTop>
-        </React.Fragment>
-    );
-}
+export default function Kanban(){
+<Container maxWidth={false}>
+    <TodoList/>
+</Container>
+<IconButton color="primary" onClick={handleNewTodo}>
+    <AddIcon />
+</IconButton>
+<NewTodo ref={newTodoRef} />}
