+++ frontend/src/components/MenuScrollToTopContainer/MenuScrollToTopContainer.js
@@ -1,0 +1,87 @@
+import * as React from 'react';
+import PropTypes from 'prop-types';
+import AppBar from '@mui/material/AppBar';
+import Toolbar from '@mui/material/Toolbar';
+import Typography from '@mui/material/Typography';
+import CssBaseline from '@mui/material/CssBaseline';
+import useScrollTrigger from '@mui/material/useScrollTrigger';
+import Box from '@mui/material/Box';
+import Container from '@mui/material/Container';
+import Fab from '@mui/material/Fab';
+import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp';
+import Zoom from '@mui/material/Zoom';
+import TodoList from "../TodoList/TodoList";
+import {IconButton} from "@mui/material";
+import AddIcon from '@mui/icons-material/Add';
+import NewTodo from '../NewTodo/NewTodo'
+import {useRef} from "react";
+
+
+function ScrollTop(props) {
+    const { children } = props;
+    const trigger = useScrollTrigger({
+        disableHysteresis: true,
+        threshold: 100,
+    });
+
+    const handleClick = (event) => {
+        const anchor = (event.target.ownerDocument || document).querySelector(
+            '#back-to-top-anchor',
+        );
+
+        if (anchor) {
+            anchor.scrollIntoView({
+                behavior: 'smooth',
+                block: 'center',
+            });
+        }
+    };
+
+    return (
+        <Zoom in={trigger}>
+            <Box
+                onClick={handleClick}
+                role="presentation"
+                sx={{  position: 'fixed', bottom: 16, right: 16 }}
+            >
+                {children}
+            </Box>
+        </Zoom>
+    );
+}
+
+ScrollTop.propTypes = {
+    children: PropTypes.element.isRequired,
+
+};
+
+
+export default function BackToTop(props) {
+    const anchorRef = useRef(null);
+
+
+    function handleNewTodo(){
+    newTodoRef.current.handleClick(anchorRef)
+    }
+
+    const newTodoRef = useRef();
+    return (
+        <React.Fragment>
+            <CssBaseline />
+            <AppBar ref={anchorRef}>
+                <Toolbar>
+                    <Typography variant="h6" component="div">
+                        Scroll to see button
+                    </Typography>
+                </Toolbar>
+            </AppBar>
+            <Toolbar id="back-to-top-anchor" />
+
+            <ScrollTop {...props}>
+                <Fab color="secondary" size="small" aria-label="scroll back to top">
+                    <KeyboardArrowUpIcon />
+                </Fab>
+            </ScrollTop>
+        </React.Fragment>
+    );
+}
