+++ frontend/src/components/NewTodo/NewTodo.js
@@ -3,10 +3,14 @@ import Box from '@mui/material/Box';
 import Popper from '@mui/material/Popper';
 import {IconButton, TextField} from "@mui/material";
 import CheckIcon from '@mui/icons-material/Check';
+import {createTodo} from "../../apiService";
+import {useAppDispatch} from '../../app/hooks'
+import {getApiData} from "../TodoList/TodoListSlicer";
 
 function NewTodo(props, ref) {
     const [anchorEl, setAnchorEl] = React.useState(null);
     const [text, setText] = React.useState('');
+    const dispatch = useAppDispatch();
     useImperativeHandle(ref, () =>({
         handleClick:(f) =>{handleClick(f)}
         })
@@ -21,6 +25,8 @@ function NewTodo(props, ref) {
 
     const handleSubmit = event =>{
         event.preventDefault();
+        createTodo({description: text, status:"todo"})
+        dispatch(getApiData());
         setAnchorEl(null)
         setText('')
     }
@@ -29,15 +35,15 @@ function NewTodo(props, ref) {
     }
 
     return (
-<div>
             <Popper id={id} open={open} anchorEl={anchorEl}>
                 <Box sx={{ border: 1, p: 1, bgcolor: 'background.paper' }} component="form" onSubmit={handleSubmit}>
-                <TextField required id="todo-content" label="todo-description" value={text} onChange={(e) => handleChange(e)} />
+                <TextField required id="todo-content" label="todo-description" value={text}
+                           onChange={handleChange} />
                     <IconButton onClick={handleSubmit}>
                         <CheckIcon />
                     </IconButton>
                 </Box>
-            </Popper></div>
+            </Popper>
     );
 }
 export default forwardRef(NewTodo)
