+++ frontend/src/components/TodoList/TodoList.js
@@ -3,20 +3,15 @@ import Grid from '@mui/material/Grid';
 import List from '@mui/material/List';
 import ListItem from '@mui/material/ListItem';
 import ListItemIcon from '@mui/material/ListItemIcon';
-import ListItemText from '@mui/material/ListItemText';
 import Checkbox from '@mui/material/Checkbox';
 import Button from '@mui/material/Button';
 import Paper from '@mui/material/Paper';
 import Todo from "../Todo/Todo";
-import styled from 'styled-components'
 
 import {useAppDispatch, useAppSelector} from '../../app/hooks';
-import {getAllTodos,getApiData, selectGetAllTodos, selectGetAllDoing, selectGetAllDone} from './TodoListSlicer'
-import {createTodo, readTodos, updateTodo, deleteTodo} from '../../apiService'
+import {getApiData, selectGetAllTodos, selectGetAllDoing, selectGetAllDone} from './TodoListSlicer'
+import { updateTodo,} from '../../apiService'
 
-function not(a, b) {
-    return a.filter((value) => b.indexOf(value) === -1);
-}
 
 function intersection(a, b) {
     return a.filter((value) => b.indexOf(value) !== -1);
@@ -25,10 +20,6 @@ function intersection(a, b) {
 export default function TodoList() {
     const dispatch = useAppDispatch();
     const [checked, setChecked] = React.useState([]);
-    // const [left, setLeft] = React.useState([0, 1, 2, 3]);
-    // const [middle, setMiddle] = React.useState([4, 5, 6, 7]);
-    // const [right, setRight] = React.useState([8, 9, 10, 11]);
-
 
     React.useEffect(() => {
         dispatch(getApiData())
@@ -50,68 +41,54 @@ export default function TodoList() {
         } else {
             newChecked.splice(currentIndex, 1);
         }
-
         setChecked(newChecked);
     };
 
-    // const handleAllRight = () => {
-    //     setMiddle(middle.concat(left));
-    //     setLeft([]);
-    // };
-
     const advanceTodo = (status) => {
         status === 'doing' ?
-            leftChecked.forEach(todo => updateTodo(todo.id, {...todo, status:'doing'}).then(()=>dispatch(getApiData()))) :
-            middleChecked.forEach(todo => updateTodo(todo.id, {...todo, status:'done'}).then(()=>dispatch(getApiData())));
-        // dispatch(getApiData())
-        // setMiddle(middle.concat(leftChecked));
-        // setLeft(not(left, leftChecked));
-        // setChecked(not(checked, leftChecked));
+            leftChecked.forEach(todo => updateTodo(todo.id, {...todo, status:'doing'})
+                .then(()=>dispatch(getApiData()))) :
+            middleChecked.forEach(todo => updateTodo(todo.id, {...todo, status:'done'})
+                .then(()=>dispatch(getApiData())));
     };
 
     const revertTodo = (status) => {
         status === 'doing' ?
             middleChecked.forEach(todo => updateTodo(todo.id, {...todo, status:'todo'}).then(()=>dispatch(getApiData()))) :
             rightChecked.forEach(todo => updateTodo(todo.id, {...todo, status:'doing'}).then(()=>dispatch(getApiData())));
-        // dispatch(getApiData())
-
-        // setLeft(left.concat(middleChecked));
-        // setMiddle(not(middle, middleChecked));
-        // setChecked(not(checked, middleChecked));
     };
 
-    // const handleAllLeft = () => {
-    //     setLeft(left.concat(middle));
-    //     setMiddle([]);
-    // };
+    function createListItem(todo) {
+
+            const labelId = `transfer-list-item-${todo.id}-label`;
+
+            return (
+                <ListItem
+                    key={todo.id}
+                    role="listitem"
+                    button
+                    onClick={handleToggle(todo)}
+                >
+                    <ListItemIcon>
+                        <Checkbox
+                            checked={checked.indexOf(todo) !== -1}
+                            tabIndex={-1}
+                            disableRipple
+                            inputProps={{
+                                'aria-labelledby': labelId,
+                            }}
+                        />
+                    </ListItemIcon>
+                    <Todo id={todo.id} text={todo.description}/>
+                </ListItem>
+            );
+
+    }
 
     const customList = (items) => (
         <Paper sx={{ width: 400, height: 400,maxHeight:400, overflow: 'auto' }}>
             <List dense component="div" role="list">
-                {items.map((value) => {
-                    const labelId = `transfer-list-item-${value.id}-label`;
-
-                    return (
-                        <ListItem
-                            key={value.id}
-                            role="listitem"
-                            button
-                            onClick={handleToggle(value)}
-                        >
-                            <ListItemIcon>
-                                <Checkbox
-                                    checked={checked.indexOf(value) !== -1}
-                                    tabIndex={-1}
-                                    disableRipple
-                                    inputProps={{
-                                        'aria-labelledby': labelId,
-                                    }}
-                                />
-                            </ListItemIcon>
-                            <Todo id={value.id} text={value.description}/>
-                        </ListItem>
-                    );
-                })}
+                {items.map(createListItem)}
                 <ListItem />
             </List>
         </Paper>
@@ -144,33 +121,13 @@ export default function TodoList() {
                     >
                         &lt;
                     </Button>
-                    {/*<Button*/}
-                    {/*    sx={{my: 0.5}}*/}
-                    {/*    variant="outlined"*/}
-                    {/*    size="small"*/}
-                    {/*    onClick={handleAllLeft}*/}
-                    {/*    disabled={middle.length === 0}*/}
-                    {/*    aria-label="move all left"*/}
-                    {/*>*/}
-                    {/*    ≪*/}
-                    {/*</Button>*/}
-                </Grid>
+                                    </Grid>
             </Grid>
             <Grid item>
                 {customList(middle)}
             </Grid>
             <Grid item>
                 <Grid container direction="column" alignItems="center">
-                    {/*<Button*/}
-                    {/*    sx={{my: 0.5}}*/}
-                    {/*    variant="outlined"*/}
-                    {/*    size="small"*/}
-                    {/*    // onClick={handleAllRight}*/}
-                    {/*    disabled={left.length === 0}*/}
-                    {/*    aria-label="move all right"*/}
-                    {/*>*/}
-                    {/*    ≫*/}
-                    {/*</Button>*/}
                     <Button
                         sx={{my: 0.5}}
                         variant="outlined"
@@ -191,17 +148,7 @@ export default function TodoList() {
                     >
                         &lt;
                     </Button>
-                    {/*<Button*/}
-                    {/*    sx={{my: 0.5}}*/}
-                    {/*    variant="outlined"*/}
-                    {/*    size="small"*/}
-                    {/*    onClick={handleAllLeft}*/}
-                    {/*    disabled={middle.length === 0}*/}
-                    {/*    aria-label="move all left"*/}
-                    {/*>*/}
-                    {/*    ≪*/}
-                    {/*</Button>*/}
-                </Grid>
+                                    </Grid>
             </Grid>
             <Grid item>
                 {customList(right)}
