+++ frontend/src/components/TodoList/TodoListSlicer.js
@@ -19,22 +19,7 @@ export const getApiData = createAsyncThunk(
 export const todoListSlice = createSlice({
     name:'todoList',
     initialState,
-    // reducers:{
-    //     getAllTodos:(state) => {
-    //         state.todos =[123]
-    //         console.log(state)
-    //        readTodos().then(result => {
-    //            console.log(result)
-    //            console.log(state)
-    //            // state.todos.concat([1,2,3])
-    //            // if(result)state.todos = [...result]
-    //            // state.todos = [...result.filter(todo => {todo.status === 'todo'})]
-    //            // state.doing = result.filter(todo => todo.status === 'doing')
-    //            // state.done = result.filter(todo => todo.status === 'done')
-    //        })
-    //     }
-    // }
-    extraReducers:(builder) => {
+      extraReducers:(builder) => {
         builder
             .addCase(getApiData.pending, (state) =>{
